starter-project/llm-sec-final-project-starter/requirements.txt
google-genai>=1.0.0
python-dotenv>=1.0.1
pydantic>=2.9.0
pyyaml>=6.0.2
scikit-learn>=1.3.0
rich>=13.8.0
